Update EventEmitter.h
Using from different projects in the same xcode workspace causes typeid returns different type_info for the same type (text name of the type is the same but its type_index is different). Since name is the same, using std::string as key of listeners map fixes the issue.
One more change - I'm not sure whether this is a typo or not, but you use type (not Type) on line 52, I fixed this as well. If this is incorrect please revert.